Testing a tiny little change that organizes the script better.
This will be unnoticable to the user, but to reduce copy/paste, I made a
little label to unify Choice1's options and signal its end. We miiight
want to come up with a naming convention for these labels because
"Choice1" will be pretty vague when you're specifically searching for
"The scene where BE girl asks if MC likes fake or real boobs" or
something.