Splash Screen and Free Choice for Introductions
This includes an update with a splash-screen to display the
ExpansionGames logo before the main menu is shown; also includes the
groundwork for designing a means to allow the player to select which
character intro they'd like to view, in whichever order they want. Some
writing tweaks are still needed, in the future, to ensure this new
system doesn't lead to continuity errors in the current script.