Ren'Py Format Update
So thanks to molotovchainsaw for helping me locate this bug, that with a
new update of Ren'Py's software. All image tags for the character art
cannot have hyphens separating the acronym (AE, BE, PRG, etc.) from the
specifying marker (1a, SC-1, 2a, etc.) so that it can clearly see that
the acronym (what we define the character as) is the tag to associate to
each image. This will prevent multiple art of the same character from
appearing on screen unintentionally.

Furthermore, due to some relocation of files, it is not recommended to
directly open the .exe of GrowthAcademy in order for the Python
libraries to match up and the game to function. Rather, if you are
editing the code directly for distribution building, please open the
game through the Ren'Py launcher when testing out new code