Added AE027 + new method
New method determines which is the second highest girl in terms of affection and returns the result. The parameter is the girl which it should ignore during the check.

It could be written differently, where it instead returns an array (but sorted on affection). But figured it wont be necessary for the coming time.