WG085 and ManualTrack Update
Includes a bugfix for text in BE048 as well as new song cues; primarily removing the requirement to have heard a piece of music before it is present in the music room. I will run this by staff on if they still want to move forward with this, in next build.